backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Azure Services
azure-ai-formrecognizer>=3.3.2
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
openai>=1.3.7

# Document Processing
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
PyPDF2>=3.0.1
Pillow>=10.1.0

# Database & Auth
sqlalchemy>=2.0.23
alembic>=1.12.1
pymysql>=1.1.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
aiofiles>=23.2.1
httpx>=0.25.2

# Text Processing
tiktoken>=0.5.2
langchain>=0.0.340
langchain-community>=0.0.6
